게시물 삭제 시 권한체크

+++ src/main/java/com/sbs/example/lolHi/controller/usr/ArticleController.java
@@ -80,6 +80,15 @@ public class ArticleController {
 			return "common/redirect";		
 		}
 		
+		Article article = articleService.getArticleById(id);
+		
+		if (article.getMemberId() != loginedMemberId) {
+			model.addAttribute("msg", "권한이 없습니다.");
+			model.addAttribute("historyBack", true);
+			
+			return "common/redirect";	
+		}
+		
 		articleService.deleteArticle(id);	
 		
 		model.addAttribute("msg", String.format("%d 글이 삭제되었습니다.", id));
